Move channel tests out of source modules
Run MPSC and MPMC behavior tests as integration tests through the public
API. Keep production modules free of test-only code and private
implementation access.
